refactor(auth): restructure admin routes with unified auth guard layout
