--- app/(admin)/admin/_components/admin-sidebar.tsx
@@ -1,121 +1,0 @@
-"use client";
-
-import { Button } from "@/components/ui/button";
-import { cn } from "@/lib/utils";
-import {
-  ChevronLeft,
-  ChevronRight,
-  LayoutDashboard,
-  Users,
-  FileText,
-} from "lucide-react";
-import Link from "next/link";
-import { usePathname } from "next/navigation";
-import { useState } from "react";
-
-const navigation = [
-  {
-    name: "Dashboard",
-    href: "/admin",
-    icon: LayoutDashboard,
-  },
-  {
-    name: "Models",
-    href: "/admin/models",
-    icon: Users,
-  },
-  {
-    name: "Form Submissions",
-    href: "/admin/form-submissions",
-    icon: FileText,
-  },
-];
-
-export function AdminSidebar() {
-  const pathname = usePathname();
-  const [isCollapsed, setIsCollapsed] = useState(false);
-
-  return (
-    <aside
-      className={cn(
-        "flex flex-col border-r border-border bg-sidebar transition-all duration-300",
-        isCollapsed ? "w-16" : "w-64",
-      )}
-    >
-      {/* Header */}
-      <div className="flex h-16 items-center justify-between px-4 border-b border-sidebar-border">
-        {!isCollapsed && (
-          <Link href="/admin" className="flex items-center gap-2">
-            <span className="text-xl font-bold text-sidebar-foreground">
-              J.I.M.
-            </span>
-          </Link>
-        )}
-        <Button
-          variant="ghost"
-          size="icon"
-          onClick={() => setIsCollapsed(!isCollapsed)}
-          className="ml-auto text-sidebar-foreground hover:bg-sidebar-accent"
-        >
-          {isCollapsed ? (
-            <ChevronRight className="h-4 w-4" />
-          ) : (
-            <ChevronLeft className="h-4 w-4" />
-          )}
-        </Button>
-      </div>
-
-      {/* Navigation */}
-      <nav className="flex-1 space-y-1 p-2">
-        {navigation.map((item) => {
-          const isActive =
-            pathname === item.href || pathname.startsWith(item.href + "/");
-          const Icon = item.icon;
-
-          return (
-            <Link key={item.name} href={item.href}>
-              <div
-                className={cn(
-                  "flex items-center gap-3 rounded-lg px-3 py-2 text-sm font-medium transition-colors",
-                  isActive
-                    ? "bg-sidebar-accent text-sidebar-accent-foreground"
-                    : "text-sidebar-foreground hover:bg-sidebar-accent/50 hover:text-sidebar-accent-foreground",
-                )}
-              >
-                <Icon className="h-5 w-5 flex-shrink-0" />
-                {!isCollapsed && <span>{item.name}</span>}
-              </div>
-            </Link>
-          );
-        })}
-      </nav>
-
-      {/* Footer */}
-      <div className="border-t border-sidebar-border p-4">
-        {!isCollapsed ? (
-          <div className="flex items-center gap-3">
-            <div className="h-8 w-8 rounded-full bg-sidebar-accent flex items-center justify-center">
-              <span className="text-sm font-medium text-sidebar-accent-foreground">
-                A
-              </span>
-            </div>
-            <div className="flex-1 min-w-0">
-              <p className="text-sm font-medium text-sidebar-foreground truncate">
-                Admin User
-              </p>
-              <p className="text-xs text-sidebar-foreground/60 truncate">
-                admin@jimmodel.com
-              </p>
-            </div>
-          </div>
-        ) : (
-          <div className="h-8 w-8 rounded-full bg-sidebar-accent flex items-center justify-center mx-auto">
-            <span className="text-sm font-medium text-sidebar-accent-foreground">
-              A
-            </span>
-          </div>
-        )}
-      </div>
-    </aside>
-  );
-}
