--- app/(auth)/layout.tsx
@@ -1,14 +1,0 @@
-import { Toaster } from "@/components/ui/sonner";
-
-export default function AuthLayout({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-  return (
-    <>
-      {children}
-      <Toaster position="top-right" />
-    </>
-  );
-}
