+++ app/admin/(admin)/@content/default.tsx
