+++ app/admin/(admin)/@content/models/(index)/_components/columns.tsx
