+++ app/admin/(admin)/@content/models/(index)/_components/models-table-skeleton.tsx
