+++ app/admin/(admin)/@content/models/[id]/basic-info/page.tsx
