+++ app/admin/(admin)/@content/models/[id]/career/page.tsx
