+++ app/admin/(admin)/@content/models/[id]/layout.tsx
