+++ app/admin/(admin)/@content/models/[id]/page.tsx
