+++ app/admin/(admin)/@content/models/[id]/physical/page.tsx
