+++ app/admin/(admin)/@content/models/[id]/portfolio-images/_components/portfolio-image-item.tsx
