+++ app/admin/(admin)/@content/models/[id]/portfolio-images/page.tsx
