+++ app/admin/(admin)/@content/models/[id]/profile-image/page.tsx
