+++ app/admin/(admin)/@content/models/[id]/status/page.tsx
