+++ app/admin/(admin)/@content/page.tsx
