+++ app/admin/(admin)/_components/admin-sidebar.tsx
@@ -1,0 +1,175 @@
+"use client";
+
+import { Button } from "@/components/ui/button";
+import {
+  DropdownMenu,
+  DropdownMenuContent,
+  DropdownMenuItem,
+  DropdownMenuSeparator,
+  DropdownMenuTrigger,
+} from "@/components/ui/dropdown-menu";
+import { authClient, useSession } from "@/lib/auth/client";
+import { cn } from "@/lib/utils";
+import {
+  ChevronLeft,
+  ChevronRight,
+  ChevronUp,
+  FileText,
+  LayoutDashboard,
+  LogOut,
+  Users,
+} from "lucide-react";
+import Link from "next/link";
+import { usePathname } from "next/navigation";
+import { useState } from "react";
+import { toast } from "sonner";
+
+const navigation = [
+  {
+    name: "Dashboard",
+    href: "/admin",
+    icon: LayoutDashboard,
+  },
+  {
+    name: "Models",
+    href: "/admin/models",
+    icon: Users,
+  },
+  {
+    name: "Form Submissions",
+    href: "/admin/form-submissions",
+    icon: FileText,
+  },
+];
+
+export function AdminSidebar() {
+  const pathname = usePathname();
+  const [isCollapsed, setIsCollapsed] = useState(false);
+  const { data: session } = useSession();
+  const [signingOut, setSigningOut] = useState(false);
+
+  const handleSignOut = async () => {
+    await authClient.signOut({
+      fetchOptions: {
+        onSuccess: () => {
+          setSigningOut(true);
+          window.location.href = "/admin/signin";
+        },
+        onError: () => {
+          setSigningOut(false);
+          window.location.href = "/admin/signin";
+        },
+        onRequest: () => {
+          toast.info("Signing out...");
+          setSigningOut(true);
+        },
+      },
+    });
+  };
+
+  return (
+    <aside
+      className={cn(
+        "flex flex-col border-r border-border bg-sidebar transition-all duration-300",
+        isCollapsed ? "w-16" : "w-64",
+      )}
+    >
+      {/* Header */}
+      <div className="flex h-16 items-center justify-between px-4 border-b border-sidebar-border">
+        {!isCollapsed && (
+          <Link href="/admin" className="flex items-center gap-2">
+            <span className="text-xl font-bold text-sidebar-foreground">
+              J.I.M.
+            </span>
+          </Link>
+        )}
+        <Button
+          variant="ghost"
+          size="icon"
+          onClick={() => setIsCollapsed(!isCollapsed)}
+          className="ml-auto text-sidebar-foreground hover:bg-sidebar-accent"
+        >
+          {isCollapsed ? (
+            <ChevronRight className="h-4 w-4" />
+          ) : (
+            <ChevronLeft className="h-4 w-4" />
+          )}
+        </Button>
+      </div>
+
+      {/* Navigation */}
+      <nav className="flex-1 space-y-1 p-2">
+        {navigation.map((item) => {
+          const isActive =
+            pathname === item.href || pathname.startsWith(item.href + "/");
+          const Icon = item.icon;
+
+          return (
+            <Link key={item.name} href={item.href}>
+              <div
+                className={cn(
+                  "flex items-center gap-3 rounded-lg px-3 py-2 text-sm font-medium transition-colors",
+                  isActive
+                    ? "bg-sidebar-accent text-sidebar-accent-foreground"
+                    : "text-sidebar-foreground hover:bg-sidebar-accent/50 hover:text-sidebar-accent-foreground",
+                )}
+              >
+                <Icon className="h-5 w-5 flex-shrink-0" />
+                {!isCollapsed && <span>{item.name}</span>}
+              </div>
+            </Link>
+          );
+        })}
+      </nav>
+
+      {/* Footer */}
+      <div className="border-t border-sidebar-border px-4 py-2">
+        {!isCollapsed ? (
+          <DropdownMenu>
+            <DropdownMenuTrigger asChild>
+              <Button
+                variant="ghost"
+                className="w-full justify-between h-auto p-2 hover:bg-sidebar-accent"
+              >
+                <p className="text-sm font-medium text-sidebar-foreground truncate flex-1 text-left">
+                  {session?.user?.name || "Loading..."}
+                </p>
+                <ChevronUp className="h-4 w-4 text-sidebar-foreground/60 flex-shrink-0" />
+              </Button>
+            </DropdownMenuTrigger>
+            <DropdownMenuContent align="end" side="top" className="w-56">
+              <DropdownMenuItem onClick={handleSignOut}>
+                <LogOut className="mr-2 h-4 w-4" />
+                Sign out
+              </DropdownMenuItem>
+            </DropdownMenuContent>
+          </DropdownMenu>
+        ) : (
+          <DropdownMenu>
+            <DropdownMenuTrigger asChild>
+              <Button
+                variant="ghost"
+                size="icon"
+                className="mx-auto hover:bg-sidebar-accent"
+              >
+                <ChevronUp className="h-4 w-4 text-sidebar-foreground/60" />
+              </Button>
+            </DropdownMenuTrigger>
+            <DropdownMenuContent align="end" side="right" className="w-56">
+              <div className="px-2 py-1.5">
+                <p className="text-sm font-medium text-sidebar-foreground truncate">
+                  {session?.user?.name || "Loading..."}
+                </p>
+              </div>
+              <DropdownMenuSeparator />
+              <DropdownMenuItem onClick={handleSignOut}>
+                <LogOut className="mr-2 h-4 w-4" />
+                Sign out
+              </DropdownMenuItem>
+            </DropdownMenuContent>
+          </DropdownMenu>
+        )}
+      </div>
+    </aside>
+  );
+}
