+++ package.json
@@ -14,7 +14,8 @@
     "db:studio": "drizzle-kit studio",
     "seed": "tsx scripts/seed-models.ts",
     "seed:cleanup": "tsx scripts/cleanup-seeded-models.ts",
-    "db:clear": "tsx scripts/clear-all-models.ts"
+    "db:clear": "tsx scripts/clear-all-models.ts",
+    "create-user": "tsx scripts/create-user.ts"
   },
   "dependencies": {
     "@hono/zod-validator": "^0.7.4",
